make `build_deps` a stand alone function
